Refactor audio type imports and restructure audio types for better organization
