+++ src/audio/AudioManager.ts
@@ -1,5 +1,5 @@
-import { IAudioManager, MusicTrackId, SoundEffectId } from "../types/audio";
 import { AudioAsset, AudioConfig } from "./";
+import { IAudioManager, MusicTrackId, SoundEffectId } from "./types";
 
 export class AudioManager implements IAudioManager {
   private _masterVolume: number = 1.0;
